Move form to attendance
